[client] rework deprecation strategy
