+++ pycti/entities/stixCyberObservables/opencti_stix_cyber_observable_properties.py
